# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sediba Aesthetic & Wellness Clinic
servers:
  - url: /api
    description: Base API path
...
paths:
  /admin/appointments/{id}/refunds:
    post:
      operationId: recordAppointmentRefund
      tags: [admin]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentRefundInput" }
      responses:
        "200":
          description: Externally issued refund recorded against an original payment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppointmentReceipts" }
...
  /admin/appointments/{id}/receipts:
    get:
      operationId: getAppointmentReceipts
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Appointment payment balance and history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppointmentReceipts" }
    post:
      operationId: recordAppointmentReceipt
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentReceiptInput" }
      responses:
        "200":
          description: Recorded in-salon receipt and updated booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppointmentReceipts" }
...
components:
  schemas:
...
    AppointmentReceiptInput:
      type: object
      required: [requestId, amountCents, method, reference, received]
      properties:
        requestId: { type: string, minLength: 16, maxLength: 100 }
        amountCents: { type: integer, minimum: 1, maximum: 100000000 }
        method: { type: string, enum: [cash, eft, card_external] }
        reference: { type: string, minLength: 1, maxLength: 1000 }
        received: { type: boolean }
    AppointmentReceipt:
      type: object
      required: [id, amountCents, method, reference, createdAt, refundedCents, refundableCents]
      properties:
        id: { type: integer }
        amountCents: { type: integer }
        method: { type: string }
        reference: { type: string }
        createdAt: { type: string }
        refundedCents: { type: integer }
        refundableCents: { type: integer }
    AppointmentRefundInput:
      type: object
      required: [requestId, paymentId, amountCents, method, reference, reason, returned]
      properties:
        requestId: { type: string, minLength: 16, maxLength: 100 }
        paymentId: { type: integer, minimum: 1 }
        amountCents: { type: integer, minimum: 1, maximum: 100000000 }
        method: { type: string, enum: [cash, eft, card_external, yoco] }
        reference: { type: string, minLength: 1, maxLength: 200 }
        reason: { type: string, minLength: 1, maxLength: 1000 }
        returned: { type: boolean }
    AppointmentRefund:
      type: object
      required: [id, paymentId, amountCents, method, reference, reason, createdAt]
      properties:
        id: { type: integer }
        paymentId: { type: integer }
        amountCents: { type: integer }
        method: { type: string }
        reference: { type: string }
        reason: { type: string }
        createdAt: { type: string }
    AppointmentReceipts:
      type: object
      required: [appointmentId, bookingStatus, totalCents, paidCents, outstandingCents, inSalon, blockedReason, receipts, refundedCents, netReceiptsCents, refunds]
      properties:
        appointmentId: { type: integer }
        bookingStatus: { type: string }
        totalCents: { type: integer }
        paidCents: { type: integer }
        outstandingCents: { type: integer }
        inSalon: { type: boolean }
        blockedReason: { type: string }
        refundedCents: { type: integer }
        netReceiptsCents: { type: integer }
        refunds:
          type: array
          items: { $ref: "#/components/schemas/AppointmentRefund" }
        receipts:
          type: array
          items: { $ref: "#/components/schemas/AppointmentReceipt" }